Update DATAIN
Now the data from the csv file is converted into a dictionary where the keys are the headers, and the value is a list of all the values under the header. The first index in the list corresponds to the value for the first row of the data and so on.