Added finished cleaning file.
The finished cleaning file removes the 11% of visits that visit an unscraped URL, and then also removes any visits whose path length is <=3 and whose time is < 20 seconds. This reduces the number of visits, and therefore rows, in the date from ~6,100,000 to 3,583,310. We must motivate why we removed this amount of data. Personally, this increases the quality of the dataset, and for our algorithm, the information scent model, is also important.